requirements.txt
supabase==2.6.0      
streamlit==1.39.0  
python-dotenv
